src/ai_agent_hybrid/requirements.txt
# Core dependencies
python-dotenv>=1.0.0
asyncio>=3.4.3

# MCP Protocol
mcp>=0.9.0

# Google Gemini & ADK
google-generativeai>=0.8.0
google-adk>=0.1.0

# Database
psycopg2-binary>=2.9.0

# API clients
aiohttp>=3.9.0
requests>=2.31.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization (for charts)
matplotlib>=3.7.0
mplfinance>=0.12.0
Pillow>=10.0.0

# Stock market data
vnstock>=0.2.0

# Validation
pydantic>=2.5.0

# Discord bot (optional)
discord.py>=2.3.0

# Web API (optional)
fastapi>=0.108.0
uvicorn>=0.25.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Type checking
mypy>=1.7.0

# Code quality
black>=23.0.0
flake8>=6.1.0
